Define segment_domain_from_email in its own file
